Renamed simulation_communication and aggregator_communication under redis_connections package to simulation and aggregator respectively.
